refactor: auth 디렉토리 삭제 및 Dto 파일 통합
- auth 디렉토리를 삭제하고 모든 Dto들을 하나의 auth.ts 파일로 통합
- AuthUserDto를 SignUpDto로 이름 변경 (회원가입에 사용되는 Dto로 명확히 구분)
- import 경로 수정 및 코드 구조 간소화
