+++ src/interfaces/auth.ts
@@ -6,3 +6,10 @@ export interface LoginUserDto {
 export interface LoginServiceDto extends LoginUserDto {
   userId: string;
 }
+
+export interface SignUpDto {
+  email: string;
+  password?: string;
+  nickname: string;
+  provider?: string;
+}
