--- src/interfaces/auth/authUser.ts
@@ -1,6 +1,0 @@
-export interface AuthUserDto {
-  email: string;
-  password?: string;
-  nickname: string;
-  provider?: string;
-}
