+++ src/routes/auth.ts
@@ -10,8 +10,7 @@ import {
   BasicResponse,
   MultipleUserDataResponse
 } from '../interfaces/response';
-import { AuthUserDto } from '../interfaces/auth/authUser';
-import { LoginUserDto, LoginServiceDto } from '../interfaces/auth/loginUser';
+import { LoginUserDto, LoginServiceDto, SignUpDto } from '../interfaces/auth';
 import { validate } from '../middleware/express-validation';
 import { body, header } from 'express-validator';
 
@@ -169,7 +168,7 @@ authRouter.post(
   ]),
   async (req: Request, res: Response) => {
     try {
-      const newUser: AuthUserDto = req.body;
+      const newUser: SignUpDto = req.body;
       const result: BasicResponse = await AuthService.saveUser(newUser);
 
       if (result.result === true) {
